api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Thalida Noel <homely@thalida.com>"]
license = "GNU AFFERO GENERAL PUBLIC LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.1"
requests = "^2.30.0"
beautifulsoup4 = "^4.12.2"
dj-database-url = "^2.0.0"
django-unfold = "^0.6.3"
psycopg2-binary = "^2.9.6"
django-cors-headers = "^4.0.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.2"
whitenoise = {extras = ["brotli"], version = "^6.4.0"}
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
drf-spectacular = "^0.26.2"
drf-social-oauth2 = "^2.1.2"
oauth2-provider = "^0.0"


[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
httpx = "^0.24.0"
pylint-django = "^2.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
